refactor: move walkthrough types to /types/walkthrough
