+++ components/walkthroughs/Main/ParticipantsList.tsx
@@ -3,7 +3,7 @@ import Seller from "./Seller";
 import BuyerLost from "./BuyerLost";
 import SellerLost from "./SellerLost";
 
-import { Data, Seller as SellerType, Buyer as BuyerType } from "@/types/index";
+import { Data, Seller as SellerType, Buyer as BuyerType } from "@/types/walkthrough";
 
 type Props = {
   buyers: BuyerType[];
