+++ components/walkthroughs/sidebar/Description.tsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
 
-import { Data } from "@/types/index";
+import { Data } from "@/types/walkthrough";
 import { fadeIn } from "@/utils/animations";
 import { ReactNode } from "react";
